ml_service/requirements.txt
flask
pandas
numpy
scikit-learn
joblib
nltk
flask-cors
python-whois
requests
